Added support for High Sierra (10.13) using grahamperrin gist from https://github.com/geerlingguy/macos-virtualbox-vm/issues/28

+++ prepare-iso.sh
@@ -1,10 +1,9 @@
 #!/bin/bash
-
 #
-# This script will create a bootable ISO image from the installer application for El Capitan (10.11) or the new Sierra (10.12) macOS.
-# Restructured a bit, and adapted the 10.11 script from this URL:
-# https://forums.virtualbox.org/viewtopic.php?f=22&t=77068&p=358865&hilit=elCapitan+iso#p358865
-#
+# This script will create a bootable ISO image from the installer application for
+# - El Capitan (10.11)
+# - Sierra (10.12) macOS
+# - High Sierra (10.13) macOS
 
 #
 # createISO
@@ -62,24 +61,42 @@ function createISO()
     echo
     echo Restore the Base System into the ${isoName} ISO image
     echo --------------------------------------------------------------------------
-    echo $ asr restore -source /Volumes/install_app/BaseSystem.dmg -target /Volumes/install_build -noprompt -noverify -erase
-    asr restore -source /Volumes/install_app/BaseSystem.dmg -target /Volumes/install_build -noprompt -noverify -erase
+    if [ "${isoName}" == "HighSierra" ] ; then
+		  echo $ time asr restore -source "${installerAppName}"/Contents/SharedSupport/BaseSystem.dmg -target /Volumes/install_build -noprompt -noverify -erase
+    	time asr restore -source "${installerAppName}"/Contents/SharedSupport/BaseSystem.dmg -target /Volumes/install_build -noprompt -noverify -erase
+	  else
+		  echo $ asr restore -source /Volumes/install_app/BaseSystem.dmg -target /Volumes/install_build -noprompt -noverify -erase
+    	asr restore -source /Volumes/install_app/BaseSystem.dmg -target /Volumes/install_build -noprompt -noverify -erase
+	  fi
 
     echo
     echo Remove Package link and replace with actual files
     echo --------------------------------------------------------------------------
-    echo $ rm /Volumes/OS\ X\ Base\ System/System/Installation/Packages
-    rm /Volumes/OS\ X\ Base\ System/System/Installation/Packages
-    echo $ cp -Rp /Volumes/install_app/Packages /Volumes/OS\ X\ Base\ System/System/Installation/
-    cp -Rp /Volumes/install_app/Packages /Volumes/OS\ X\ Base\ System/System/Installation/
+
+	  if [ "${isoName}" == "HighSierra" ] ; then
+	    echo $ time ditto -V /Volumes/install_app/Packages /Volumes/OS\ X\ Base\ System/System/Installation/
+    	time ditto -V /Volumes/install_app/Packages /Volumes/OS\ X\ Base\ System/System/Installation/
+	  else
+    	echo $ rm /Volumes/OS\ X\ Base\ System/System/Installation/Packages
+    	rm /Volumes/OS\ X\ Base\ System/System/Installation/Packages
+    	echo $ cp -Rp /Volumes/install_app/Packages /Volumes/OS\ X\ Base\ System/System/Installation/
+    	cp -Rp /Volumes/install_app/Packages /Volumes/OS\ X\ Base\ System/System/Installation/
+	  fi
 
     echo
     echo Copy macOS ${isoName} installer dependencies
     echo --------------------------------------------------------------------------
-    echo $ cp -Rp /Volumes/install_app/BaseSystem.chunklist /Volumes/OS\ X\ Base\ System/BaseSystem.chunklist
-    cp -Rp /Volumes/install_app/BaseSystem.chunklist /Volumes/OS\ X\ Base\ System/BaseSystem.chunklist
-    echo $ cp -Rp /Volumes/install_app/BaseSystem.dmg /Volumes/OS\ X\ Base\ System/BaseSystem.dmg
-    cp -Rp /Volumes/install_app/BaseSystem.dmg /Volumes/OS\ X\ Base\ System/BaseSystem.dmg
+	  if [ "${isoName}" == "HighSierra" ] ; then
+	    echo $ ditto -V "${installerAppName}"/Contents/SharedSupport/BaseSystem.chunklist /Volumes/OS\ X\ Base\ System/BaseSystem.chunklist
+    	ditto -V "${installerAppName}"/Contents/SharedSupport/BaseSystem.chunklist /Volumes/OS\ X\ Base\ System/BaseSystem.chunklist
+    	echo $ time ditto -V "${installerAppName}"/Contents/SharedSupport/BaseSystem.dmg /Volumes/OS\ X\ Base\ System/BaseSystem.dmg
+    	time ditto -V "${installerAppName}"/Contents/SharedSupport/BaseSystem.dmg /Volumes/OS\ X\ Base\ System/BaseSystem.dmg	
+	  else
+    	echo $ cp -Rp /Volumes/install_app/BaseSystem.chunklist /Volumes/OS\ X\ Base\ System/BaseSystem.chunklist
+    	cp -Rp /Volumes/install_app/BaseSystem.chunklist /Volumes/OS\ X\ Base\ System/BaseSystem.chunklist
+    	echo $ cp -Rp /Volumes/install_app/BaseSystem.dmg /Volumes/OS\ X\ Base\ System/BaseSystem.dmg
+    	cp -Rp /Volumes/install_app/BaseSystem.dmg /Volumes/OS\ X\ Base\ System/BaseSystem.dmg
+	  fi
 
     echo
     echo Unmount the installer image
@@ -100,10 +117,10 @@ function createISO()
     hdiutil resize -size `hdiutil resize -limits /tmp/${isoName}.sparseimage | tail -n 1 | awk '{ print $1 }'`b /tmp/${isoName}.sparseimage
 
     echo
-    echo Convert the sparse bundle to ISO/CD master
+    echo Convert ${isoName} the sparse bundle to ISO/CD master
     echo --------------------------------------------------------------------------
-    echo $ hdiutil convert /tmp/${isoName}.sparseimage -format UDTO -o /tmp/${isoName}
-    hdiutil convert /tmp/${isoName}.sparseimage -format UDTO -o /tmp/${isoName}
+    echo $ time hdiutil convert /tmp/${isoName}.sparseimage -format UDTO -o /tmp/${isoName}
+    time hdiutil convert /tmp/${isoName}.sparseimage -format UDTO -o /tmp/${isoName}
 
     echo
     echo Remove the sparse bundle
@@ -140,7 +157,6 @@ function installerExists()
 # Main script code
 #
 # Eject installer disk in case it was opened after download from App Store
-# added support for multiple disks
 for disk in $(hdiutil info | grep /dev/disk | grep partition | cut -f 1); do
   hdiutil detach -force ${disk}
 done
@@ -148,22 +164,28 @@ done
 # See if we can find either the ElCapitan or the Sierra installer.
 # If successful, then create the iso file from the installer.
 
-installerExists "Install macOS Sierra.app"
+installerExists "Install macOS High Sierra.app"
 result=$?
 if [ ${result} -eq 0 ] ; then
-  createISO "Install macOS Sierra.app" "Sierra"
+  createISO "Install macOS High Sierra.app" "HighSierra"
 else
-  installerExists "Install OS X El Capitan.app"
+  installerExists "Install macOS Sierra.app"
   result=$?
   if [ ${result} -eq 0 ] ; then
-    createISO "Install OS X El Capitan.app" "ElCapitan"
+    createISO "Install macOS Sierra.app" "Sierra"
   else
-    installerExists "Install OS X Yosemite.app"
+    installerExists "Install OS X El Capitan.app"
     result=$?
     if [ ${result} -eq 0 ] ; then
-      createISO "Install OS X Yosemite.app" "Yosemite"
+      createISO "Install OS X El Capitan.app" "ElCapitan"
     else
-      echo "Could not find installer for Yosemite (10.10), El Capitan (10.11) or Sierra (10.12)."
+      installerExists "Install OS X Yosemite.app"
+      result=$?
+      if [ ${result} -eq 0 ] ; then
+        createISO "Install OS X Yosemite.app" "Yosemite"
+      else
+        echo "Could not find installer for Yosemite (10.10), El Capitan (10.11), Sierra (10.12) or High Sierra (10.13)."
+      fi
     fi
   fi
 fi
